Pipe-able, vectorized, and lightweight implementation of git in R

Apply a Git command to each path.

walk_git(path, command, verbose = FALSE, stop_on_error = TRUE, ...)

Arguments

path
Path to one or multiple Git repos.

command
A Git command, e.g. "status" or "log --oneline -n 1".

verbose
Print Git's output?

stop_on_error
If Git fails, do you want an R error?

...
Other arguments passed to system.

Value

walk_git() is called for its side effect; it returns path invisibly.

Examples

library(magrittr)

# helper
walk <- function(x, f, ...) {
  lapply(x, f, ...)
  invisible(x)
}

repos <- file.path(tempdir(), paste0("repo", 1:2))
repos %>% walk(dir.create)

# Fails because the repo isn't initialized
repos %>%
  walk_git("status") %>%
  try()

#> Error : fatal: not a git repository (or any of the parent directories): .git

# Don't throw an error
repos %>%
  walk_git("status", stop_on_error = FALSE)

repos %>% walk_git("init")

repos %>% walk_git("status")
repos %>% walk_git("status", verbose = TRUE)

#> $`/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpCxAlj9/repo1`
#> [1] "On branch master"  ""                  "No commits yet"   
#> [4] ""                  "nothing to commit"
#> 
#> $`/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpCxAlj9/repo2`
#> [1] "On branch master"  ""                  "No commits yet"   
#> [4] ""                  "nothing to commit"
#> 

repos %>%
  walk_git("add .") %>%
  walk_git("commit -m 'Initialize' --allow-empty") %>%
  walk_git("log --oneline -n 1", verbose = TRUE)

#> $`/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpCxAlj9/repo1`
#> [1] "c8225da Initialize"
#> 
#> $`/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T//RtmpCxAlj9/repo2`
#> [1] "c8225da Initialize"
#> 

# Cleanup
walk(repos, unlink, recursive = TRUE)